MCOL-1822 Intermediate checkin. DISTINCT not working.

+++ utils/funcexp/functor_str.h
@@ -57,7 +57,7 @@ public:
         return strtod(getStrVal(row, fp, isNull, op_ct).c_str(), NULL);
     }
 
-    double getLongDoubleVal(rowgroup::Row& row,
+    long double getLongDoubleVal(rowgroup::Row& row,
                         FunctionParm& fp,
                         bool& isNull,
                         execplan::CalpontSystemCatalog::ColType& op_ct)
